chore: entity 추가 수정

+++ src/main/java/com/book_everywhere/domain/tagged/TaggedRepository.java
@@ -2,5 +2,5 @@ package com.book_everywhere.domain.tagged;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaggedRepository extends JpaRepository<Tagged,Integer> {
+public interface TaggedRepository extends JpaRepository<Tagged,Long> {
 }
